brains/Physics2.brain/outputs/code-creation/agent_1766440379610_hhonegh/requirements.txt
# Pinned runtime dependencies for deterministic installs.
# These are required to run the repository tooling and scripts on clean environments.

PyYAML==6.0.2
